requirements.txt
appdirs==1.4.4
black==20.8b1
click==7.1.2
cycler==0.10.0
glfw==2.0.0
kiwisolver==1.3.1
matplotlib==3.3.3
mypy-extensions==0.4.3
numpy==1.19.4
pathspec==0.8.1
Pillow==8.1.0
PyOpenGL==3.1.5
pyparsing==2.4.7
python-dateutil==2.8.1
regex==2020.11.13
six==1.15.0
toml==0.10.2
typed-ast==1.4.1
typing-extensions==3.7.4.3
